// frontend/src/api/generated/shop/schemas/supplier.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */

export interface Supplier {
  readonly id: number;
  /**
   * Supplier company name
   * @maxLength 100
   */
  name: string;
  /**
   * Primary contact email
   * @maxLength 254
   */
  contact_email?: string;
  /**
   * Contact phone number
   * @maxLength 20
   */
  phone?: string;
  /** Timestamp when the record was created */
  readonly created_at: string;
  /** Timestamp when the record was last updated */
  readonly updated_at: string;
}